--- src/main/kotlin/no/skatteetaten/aurora/gobo/exceptions/RestResponseException.kt
@@ -1,3 +1,0 @@
-package no.skatteetaten.aurora.gobo.exceptions
-
-class RestResponseException(message: String?, cause: Throwable? = null, code: String = "", errorMessage: String = "") : GoboException(message, cause, code, errorMessage)
