+++ src/main/kotlin/no/skatteetaten/aurora/gobo/integration/imageregistry/imageRegistryDtos.kt
@@ -2,8 +2,6 @@ package no.skatteetaten.aurora.gobo.integration.imageregistry
 
 import java.time.Instant
 
-class ImageRegistryServiceErrorException(message: String, cause: Throwable? = null) : RuntimeException(message, cause)
-
 data class ImageRepoDto(val registry: String, val namespace: String, val name: String) {
     val repository: String
         get() = listOf(registry, namespace, name).joinToString("/")
@@ -20,13 +18,15 @@ enum class ImageTagType {
 
     companion object {
         fun typeOf(tag: String): ImageTagType {
-            return if (tag.toLowerCase().equals("latest")) ImageTagType.LATEST
-            else if (tag.toLowerCase().endsWith("-snapshot")) ImageTagType.SNAPSHOT
-            else if (tag.toLowerCase().startsWith("snapshot-")) ImageTagType.AURORA_SNAPSHOT_VERSION
-            else if (tag.matches(Regex("^\\d+$"))) ImageTagType.MAJOR
-            else if (tag.matches(Regex("^\\d+.\\d+$"))) ImageTagType.MINOR
-            else if (tag.matches(Regex("^\\d+.\\d+.\\d+$"))) ImageTagType.BUGFIX
-            else ImageTagType.AURORA_VERSION
+            return when {
+                tag.toLowerCase() == "latest" -> ImageTagType.LATEST
+                tag.toLowerCase().endsWith("-snapshot") -> ImageTagType.SNAPSHOT
+                tag.toLowerCase().startsWith("snapshot-") -> ImageTagType.AURORA_SNAPSHOT_VERSION
+                tag.matches(Regex("^\\d+$")) -> ImageTagType.MAJOR
+                tag.matches(Regex("^\\d+.\\d+$")) -> ImageTagType.MINOR
+                tag.matches(Regex("^\\d+.\\d+.\\d+$")) -> ImageTagType.BUGFIX
+                else -> ImageTagType.AURORA_VERSION
+            }
         }
     }
 }
